Move `find_struct_impl` to assist utils
